# storage/api-docs/swaggernew.yaml
openapi: 3.0.0
info:
  title: 'Liftoff Card Integration API'
  description: 'Defines the interface for programmatic interaction with the ACH payment processing system. This specification details resources for managing payees (merchants), originating individual ACH transactions, administering recurring payment schedules, and retrieving real-time status updates throughout the transaction lifecycle, including confirmation of settlement or details of returns.'
  contact:
    name: 'Liftoff Card Support / Grid Funding Integration Team'
    email: support@liftoffcard.com
  version: 1.0.0
servers:
  -
    url: 'https://app.liftoffcard.com/liftoffcard-solutions-api'
    description: 'Production server'
paths:
  /api/v1/auth/token:
    post:
      tags:
        - Authentication
      summary: 'Get authentication token (JWT)'
      operationId: 212804489af549fa19ff9b679cdd33e3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: loreum_user
                password:
                  type: string
                  example: loreum_user
              type: object
      responses:
        200:
          description: 'Token generated successfully'
          content:
            application/json:
              schema:
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
                type: object
        401:
          description: 'Invalid credentials'
  /api/v1/auth/refresh:
    post:
      tags:
        - Authentication
      summary: 'Refresh an expired JWT token'
      operationId: c42c3cd7fd7fe0dbd5bcc42962059e90
      responses:
        200:
          description: 'Token refreshed successfully'
        401:
          description: 'Invalid or expired token'
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...